Run fMRI quality assurance procedure on a NIfTI data file

Run fMRI quality assurance procedure on a NIfTI data file

run_fmriqa(data_file = NULL, roi_width = 21, slice_num = NULL,
  skip = 2, tr = NULL, pix_dim = NULL, poly_det_ord = 3,
  spike_detect = FALSE, x_pos = NULL, y_pos = NULL,
  plot_title = NULL, last_vol = NULL, gen_png = TRUE,
  gen_res_csv = TRUE, gen_pdf = FALSE, gen_spec_csv = FALSE,
  png_fname = NULL, res_fname = NULL, pdf_fname = NULL,
  spec_fname = NULL, verbose = TRUE, bg_smooth = 12,
  bg_shrink = 25)

Arguments

data_file
input data in nifti format, a file chooser will open if not set

roi_width
roi analysis region in pixels (default=21)

slice_num
slice number for analysis (default=middle slice)

skip
number of initial volumes to exclude from the analysis (default=2)

tr
override the TR detected from data (seconds)

pix_dim
override the x,y,z pixel dimensions (mm) detected from data eg pixdim=c(3,3,3)

poly_det_ord
polynomial order used for detrending (default=3)

spike_detect
generate k-space spike-detection plot (default=FALSE)

x_pos
x position of ROI (default=center of gravity)

y_pos
y position of ROI (default=center of gravity)

plot_title
add a title to the png and pdf plots

last_vol
last volume number to use in the analysis

gen_png
output png plot (default=TRUE)

gen_res_csv
output csv results (default=TRUE)

gen_pdf
output pdf plot (default=FALSE)

gen_spec_csv
output csv of spectral points (default=FALSE)

png_fname
png plot filename

res_fname
csv results filename

pdf_fname
pdf plot filename

spec_fname
csv spectral data filename

verbose
provide text output while running (default=TRUE)

bg_smooth
amount to smooth background image before calculating the maximum BG percent metric (default=12mm)

bg_shrink
amount to shrink the BG image away from the object to avoid residual object signal in the maximum BG percent metric (default=25mm)

Value

dataframe of QA metrics

Examples

fname <- system.file("extdata", "qa_data.nii.gz", package = "fmriqa")
res <- run_fmriqa(data_file = fname, gen_png = FALSE, gen_res_csv = FALSE, tr = 3)
#> Reading data  : /home/runner/work/_temp/Library/fmriqa/extdata/qa_data.nii.gz
#> 
#> Basic analysis parameters
#> -------------------------
#> X,Y matrix     : 80x80
#> Slices         : 1
#> X,Y,Z pix dims : 1x1x1mm
#> TR             : 3s
#> Slice #        : 1
#> ROI width      : 21
#> Total vols     : 200
#> Analysis vols  : 198
#> 
#> QA metrics
#> ----------
#> Mean signal   : 2561.3
#> STD           : 6.4
#> Percent fluc  : 0.25
#> Drift         : 2.06
#> Drift fit     : 1.42
#> SNR           : 141.3
#> SFNR          : 137
#> RDC           : 2.84
#> TC outlier    : 2.66
#> Spec outlier  : 5.12
#> MBG percent   : 0.74